RigidBodyComponent api progress

+++ HeartEditor/src/HeartEditor/Widgets/PropertiesPanel.h
@@ -29,7 +29,7 @@ namespace Widgets
 
         bool RenderXYZSlider(Heart::HStringView8 name, f32* x, f32* y, f32* z, f32 min, f32 max, f32 step);
         void RenderScriptField(Heart::HStringView fieldName, Heart::ScriptComponent& scriptComp);
-        bool RenderCollisionChannels(Heart::HStringView8 id, u32& mask);
+        bool RenderCollisionChannels(Heart::HStringView8 id, u64& mask);
         
         // returns true if the component was deleted
         template<typename Component>
